+++ src/rv_core/instruction/executor/base/auipc.rs
@@ -1,5 +1,5 @@
 use crate::rv_core::instruction::executor::prelude::*;
 
 pub fn auipc(U { rd, imm20 }: U, x: &mut IntegerRegisters, pc: u64) {
-    x[rd] = pc + imm20 << 12;
+    x[rd] = pc + (imm20 << 12) as u64;
 }
