+++ src/rv_core/instruction/format/base.rs
@@ -10,7 +10,7 @@ pub struct R {
 pub struct I {
     pub rd: usize,
     pub rs1: usize,
-    pub imm12: u64,
+    pub imm12: i32,
 }
 
 #[derive(Clone, Debug)]
@@ -18,11 +18,11 @@ pub struct I {
 pub struct S {
     pub rs1: usize,
     pub rs2: usize,
-    pub imm12: u64,
+    pub imm12: i32,
 }
 
 #[derive(Clone, Debug)]
 pub struct U {
     pub rd: usize,
-    pub imm20: u64,
+    pub imm20: i32,
 }
